docs: readme
