+++ examples/tree/readme-custom-enumerator/main.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/charmbracelet/lipgloss/tree"
+)
+
+func main() {
+	t := tree.New(
+		"",
+		"Glossier",
+		"Claire’s Boutique",
+		tree.New("Nyx", "Foo", "Bar"),
+		"Mac",
+		"Milk",
+	).Enumerator(func(atter tree.Atter, i int, last bool) (indent string, prefix string) {
+		return "->", "->"
+	})
+	fmt.Println(t)
+}
